# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.Code:
    enum:
    - 0
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 2001
    - 2002
    - 2003
    - 2004
    - 3001
    - 3002
    - 3003
    - 3004
    - 4001
    - 4002
    - 4003
    - 4004
    type: integer
    x-enum-varnames:
    - Success
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrInvalidParams
    - ErrBindJSON
    - ErrBindQuery
    - ErrBindForm
    - ErrDatabase
    - ErrRecordNotFound
    - ErrRecordExists
    - ErrRecordInvalidate
    - ErrTokenInvalid
    - ErrTokenExpired
    - ErrTokenMissing
    - ErrPermissionDenied
...
  model.File:
    properties:
      bucket_name:
        description: 存储桶名称（OSS/S3）
        type: string
      category:
        description: '文件分类: avatar, document, image, video, other'
        type: string
      created_at:
        type: string
      extension:
        description: 文件扩展名
        type: string
      hash:
        description: 文件 Hash（MD5/SHA256）
        type: string
      height:
        description: 图片高度
        type: integer
      id:
        type: integer
      mime_type:
        description: MIME 类型
        type: string
      original_name:
        description: 原始文件名
        type: string
      path:
        description: 文件存储路径
        type: string
      saved_name:
        description: 保存的文件名（UUID）
        type: string
      size:
        description: 文件大小（字节）
        type: integer
      status:
        description: '状态: 1=正常, 2=已删除, 3=审核中'
        type: integer
      storage_type:
        description: '存储类型: local, oss, s3'
        type: string
      thumbnail_path:
        description: 可选：缩略图相关
        type: string
      thumbnail_url:
        description: 缩略图 URL
        type: string
      updated_at:
        type: string
      uploaded_by:
        description: 上传用户ID
        type: integer
      url:
        description: 访问 URL
        type: string
      width:
        description: 可选：元数据
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        $ref: '#/definitions/errors.Code'
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/cccvno1/nova
  description: |-
    Nova 是一个基于 Go + Echo 的现代化 Web 应用框架，提供用户认证、RBAC 权限管理、文件上传、异步任务队列、审计日志等功能。

    主要特性：
    - 用户认证与授权（JWT）
    - RBAC 权限管理（基于 Casbin）
    - 文件上传与管理（支持本地/OSS/S3）
    - 异步任务队列（基于 Redis）
    - 定时任务调度（基于 Cron）
    - 审计日志（自动记录所有操作）
    - 限流保护（IP/用户维度）
    - 健康检查与监控
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Nova API
  version: "1.0"
paths:
...
  /files:
    get:
      consumes:
      - application/json
      description: 获取当前用户的文件列表，支持分类过滤和分页
      parameters:
      - description: 文件分类
        enum:
        - avatar
        - document
        - image
        - video
        - audio
        - other
        in: query
        name: category
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文件列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.File'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取文件列表
      tags:
      - 文件管理
  /files/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除指定文件（逻辑删除）
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 无权限删除该文件
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除文件
      tags:
      - 文件管理
    get:
      consumes:
      - application/json
      description: 根据ID获取文件的详细信息
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文件详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.File'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取文件详情
      tags:
      - 文件管理
  /files/{id}/download:
    get:
      consumes:
      - application/json
      description: 根据文件ID下载文件，支持断点续传
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件流
          schema:
            type: file
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 下载文件
      tags:
      - 文件管理
  /files/search:
    get:
      consumes:
      - application/json
      description: 根据关键词搜索文件名，支持分页
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.File'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 搜索文件
      tags:
      - 文件管理
  /files/storage/info:
    get:
      consumes:
      - application/json
      description: 获取当前用户的存储空间使用情况统计
      produces:
      - application/json
      responses:
        "200":
          description: 存储空间信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取存储空间信息
      tags:
      - 文件管理
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件到服务器，支持图片自动生成缩略图
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      - default: other
        description: 文件分类
        enum:
        - avatar
        - document
        - image
        - video
        - audio
        - other
        in: formData
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功，返回文件信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.File'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "413":
          description: 文件过大
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件管理
  /files/upload/avatar:
    post:
      consumes:
      - multipart/form-data
      description: 上传头像文件（仅支持图片格式：jpg, jpeg, png, gif, webp）
      parameters:
      - description: 头像文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.File'
              type: object
        "400":
          description: 请求参数错误或文件格式不支持
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 上传用户头像
      tags:
      - 文件管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"